rig up options in the view
